♻️ refactor: adjust events struct
